refactor: Update farm-mp directory structure
