+++ src/pages/_mp/farms/index.tsx
@@ -1,5 +1,5 @@
-import { FarmsMpPageLayout } from 'views/Farms'
-import FarmsPage from '../farms'
+import FarmsMpPageLayout from 'views/Farms/MpPageLayout'
+import FarmsPage from '../../farms'
 
 const MpFarmsPage = () => {
   return <FarmsPage />
